Incorporate database and loosen constraint for matching Gerrit change url
Without a form of persistent storage, the LKML Gerrit Bridge has to
reupload patches to Gerrit whenever there is a restart. This problem is
resolved by the addition of a database to store patches that have been uploaded.

Additionally, numerous amounts of patches were previously failing due to
an inability to exactly match the anticipated format of a Gerrit push
response, so the constraint of this match was reduced to only search for
the success message within the Gerrit push response.